# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NetflixAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Episode": {
      "get": {
        "tags": [
          "Episode"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Episode"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Episode"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Episode"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Episode"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Episode"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Episode"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Episode"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Episode"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Episode"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Episode"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Episode"
                }
              }
            }
          }
        }
      }
    },
    "/api/Episode/{id}": {
      "get": {
        "tags": [
          "Episode"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Episode"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Episode"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Episode"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Episode"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Episode"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Episode"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Episode"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Episode"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Language": {
      "get": {
        "tags": [
          "Language"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Language"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          }
        }
      }
    },
    "/api/Language/{id}": {
      "get": {
        "tags": [
          "Language"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Language"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Language"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Episode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "contentId": {
            "type": "integer",
            "format": "int64"
          },
          "seasonId": {
            "type": "integer",
            "format": "int64"
          },
          "episodeName": {
            "type": "string",
            "nullable": true
          },
          "contentIndex": {
            "type": "integer",
            "format": "int32"
          },
          "length": {
            "type": "integer",
            "format": "int32"
          },
          "episodeQuality": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          },
          "subtitles": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Language": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "languageName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}